LogicProvider refactoring

+++ src/BalloonsPop.Common/Contracts/IRandomNumberGenerator.cs
@@ -4,6 +4,6 @@
 
     public interface IRandomNumberGenerator
     {
-        int Next(int min, int max);
+        byte Next(int min, int max);
     }
 }
